- Make DefaultStoreTest faster
  - No need to deploy any app to test store.
  - Runtime drop from 21.7 sec to 2.9 sec
- Replace the mandatory sleep in TestFrameworkTestBase to a retry on failure
- Speedup AdminTestRun by not recreating the same app jar.
  - Runtime drop from 77.6 sec to 62.1 sec
- Speedup SparkFileSetTestRun by not recreating the same app jar
  - Runtime drop from 67.3 sec to 56.3 sec
- Speedup SparkTestRun by not recreating app jar
  - Runtime drop from 169.9 to 113.1 seconds
- Speedup PreferencesHttpHandlerTest by not deploying app
  - It just need store entry, no need to run the deployment pipeline
  - Runtime drop from 18.1 to 7.1 seconds
- Speedup ServiceLifeCycleTestRun by not recreating jar
- Fixed a bug in TestBase that the applicationManager is not stopped
  if deployed with an artifact jar.